requirements.txt
bcrypt==3.1.3
cffi==1.10.0
click==6.7
Flask==0.12.2
Flask-MongoAlchemy==0.7.2
flask-mongoengine==0.9.3
Flask-PyMongo==0.5.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
MongoAlchemy==0.19
mongoengine==0.13.0
pycparser==2.18
pymongo==3.5.0
python-memcached==1.58
six==1.10.0
Werkzeug==0.12.2
WTForms==2.1
